Add zero-downtime deployment by passing ami and server_text params (in real life the ami parameter will be sufficient as all changes will be there and not in user_data but for the sack of exapmle we trigger the update of the service with server_text change. To 0-downtime deployment to work we must change 3 things: 1. Configure the name parameter of the ASG to depend directly on the name of the launch configuration. That way, each time the launch configuration changes (which it will when you update the AMI or User Data), Terraform will try to replace the ASG. 2. Set the create_before_destroy parameter of the ASG to true , so each time Terraform tries to replace it, it will create the replacement before destroying the original. 3. Set the min_elb_capacity parameter of the ASG to the min_size of the cluster so that Terraform will wait for at least that many servers from the new ASG to register in the ELB before it’ll start destroying the original ASG.
If something went wrong during the deployment, Terraform will automatically
roll back! For example, if there was a bug in v2 of your app and it failed to boot, then the Instances in the new ASG will not register with the ELB. Terraform will wait up to wait_for_capacity_timeout (default is 10 minutes) for min_elb_capacity servers of the v2
ASG to register in the ELB, after which it will consider the deployment a failure, delete the v2 ASG, and exit with an error (meanwhile, v1 of your app continues to run just fine in the original ASG).